+++ src/components/ChatInfoStatusBox.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Image, View, Text} from "react-native";
+import { Image, View, Text } from "react-native";
 import styles from "./utils/styles";
 import Pressable from "react-native/Libraries/Components/Pressable/Pressable";
 import StatusBox from "./StatusBox";
@@ -7,31 +7,29 @@ import StatusBox from "./StatusBox";
 const ChatInfoStatusOptionBox = (props) => {
   return (
     <Pressable
-        onPress={() => 
-            props.navigation.navigate(props.routeName) 
-        }
-      
-        style={({pressed}) => [
+      onPress={() =>
+        props.navigation.navigate(props.routeName, {
+          bubbleID: props.bubbleID,
+          headerTitle: props.headerTitle,
+          bubbleTitle: props.bubbleTitle,
+        })
+      }
+      style={({ pressed }) => [
         {
-            backgroundColor: pressed ? "#D9D9D9" : "#FFFFFF00",
+          backgroundColor: pressed ? "#D9D9D9" : "#FFFFFF00",
         },
-    ]}>
-
+      ]}
+    >
       <View style={styles.optionContainer}>
-        <Image
-            source = {props.icon} 
-            style={styles.chatInfoOptionIcon}
-        />
-        <Text style={styles.optionText}>
-            {props.name}
-        </Text>
+        <Image source={props.icon} style={styles.chatInfoOptionIcon} />
+        <Text style={styles.optionText}>{props.name}</Text>
       </View>
-        
+
       <View style={styles.chatInfoOptionContainer}>
         <StatusBox userStatus={props.userStatus} />
       </View>
     </Pressable>
-  )
+  );
 };
 
 export default ChatInfoStatusOptionBox;